+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version 0.2 (working)
+* Add `space_id` and `point_types` options to `points()` and `values()`
 
 ## Version 0.1 (9-Mar-2026)
 * Initial MVP